pyproject.toml
[project]
name = "recetario-whatsapp"
version = "0.1.0"
description = "Convierte chats de WhatsApp en un recetario interactivo usando IA"
authors = [
    {name = "raistln", email = "samumarfon@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.12"
keywords = ["whatsapp", "recetas", "ia", "mistral", "streamlit", "supabase"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Topic :: Text Processing",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]
dependencies = [
    "requests>=2.31.0",
    "streamlit>=1.28.0",
    "supabase>=2.0.0",
    "python-dotenv>=1.0.0",
    "mistralai>=1.0.0",
    "Pillow>=10.0.0"
]

[project.urls]
Homepage = "https://github.com/tu-usuario/recetario-whatsapp"
Repository = "https://github.com/tu-usuario/recetario-whatsapp"
Issues = "https://github.com/tu-usuario/recetario-whatsapp/issues"
Documentation = "https://github.com/tu-usuario/recetario-whatsapp/tree/main/docs"

[project.scripts]
recetario-whatsapp = "recetario_whatsapp.extractor:main"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.4.0,<8.0.0"
pytest-cov = ">=4.1.0,<5.0.0"
pytest-mock = ">=3.11.0,<4.0.0"
pytest-asyncio = ">=0.21.0,<1.0.0"
black = ">=23.0.0"
flake8 = ">=6.0.0"
mypy = ">=1.0.0"

[tool.poetry]
packages = [{include = "recetario_whatsapp", from = "src"}]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
exclude = [
    ".git",
    "__pycache__",
    "build",
    "dist",
    ".venv",
    ".pytest_cache"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
